Rearrange the furniture, vaccuum the carpet, etc.
This commit moves a bunch of methods from the Doc companion into the
class. It also renames Doc2 to Chunk, splits chunk and tree-related
functionality out of Doc into other files, and adds a bit more
documentation.
