+++ core/src/main/scala/com/github/johnynek/paiges/package.scala
@@ -1,0 +1,8 @@
+package com.github.johnynek
+
+package object paiges {
+
+  type StreamTree[T] = Stream[Either[(T, () => T), Chunk]]
+
+  type DocTree = Fix[StreamTree]
+}
