+++ core/src/test/scala/com/github/johnynek/paiges/PaigesTest.scala
@@ -122,19 +122,16 @@ the spaces""")
   }
   test("renders are constant after maxWidth") {
     forAll { (d: Doc, ws: List[Int]) =>
-      val m = Doc.maxWidth(d)
+      val m = d.maxWidth
       val maxR = d.render(m)
       val justAfter = (1 to 20).iterator
       val goodW = (justAfter ++ ws.iterator).map { w => (m + w) max m }
-      assert(goodW.forall { w =>
-        val wrender = d.render(w)
-        (wrender == maxR)
-      })
+      assert(goodW.forall { w => d.render(w) == maxR })
     }
   }
   test("if we always render the same, we compare the same") {
     forAll { (a: Doc, b: Doc) =>
-      val maxR = Doc.maxWidth(a) max Doc.maxWidth(b)
+      val maxR = a.maxWidth max b.maxWidth
       val allSame = (0 to maxR).forall { w =>
         a.render(w) == b.render(w)
       }
@@ -180,7 +177,7 @@ the spaces""")
      * flatten(c) + b.grouped == (flatten(c) + b).grouped
      */
     forAll { (b: Doc, c: Doc) =>
-      val flatC = Doc.flatten(c)
+      val flatC = c.flatten
       val left = (b.grouped + flatC)
       val right = (b + flatC).grouped
       assert((left).compare(right) == 0)
@@ -191,7 +188,7 @@ the spaces""")
   }
   test("flatten(group(a)) == flatten(a)") {
     forAll { (a: Doc) =>
-      assert(Doc.flatten(a.grouped).compare(Doc.flatten(a)) == 0)
+      assert(a.grouped.flatten.compare(a.flatten) == 0)
     }
   }
 
@@ -225,9 +222,9 @@ the spaces""")
   test("isSubDoc works correctly: group") {
     forAll { (d: Doc) =>
       import Doc._
-      val f = flatten(d)
+      val f = d.flatten
       val g = d.grouped
-      assert(isSubDoc(toDocTree(f), toDocTree(g)))
+      assert(f.isSubDocOf(g))
     }
   }
 
@@ -237,7 +234,7 @@ the spaces""")
       // we need at least 2 docs for this law
       val ds = (d0 :: d1 :: dsLong.take(4))
       val f = fill(empty, ds)
-      val g = intercalate(space, ds.map(flatten(_)))
+      val g = intercalate(space, ds.map(_.flatten))
       assert(g.isSubDocOf(f))
     }
   }
@@ -245,8 +242,8 @@ the spaces""")
   test("if isSubDoc is true, there is some width that renders the same") {
     forAll { (d1: Doc, d2: Doc) =>
       import Doc._
-      if (isSubDoc(toDocTree(d1), toDocTree(d2))) {
-        val mx = maxWidth(d1) max maxWidth(d2)
+      if (DocTree.isSubDoc(DocTree.toDocTree(d1), DocTree.toDocTree(d2))) {
+        val mx = d1.maxWidth max d2.maxWidth
         assert((0 to mx).exists { w => d1.render(w) == d2.render(w) })
       }
       else succeed
@@ -265,37 +262,37 @@ the spaces""")
   test("setDiff(a, a) == None") {
     forAll { (a: Doc) =>
       import Doc._
-      val atree = toDocTree(a)
+      val atree = DocTree.toDocTree(a)
       // we should totally empty a tree
-      assert(setDiff(atree, atree).isEmpty)
+      assert(DocTree.setDiff(atree, atree).isEmpty)
     }
   }
   test("after setDiff isSubDoc is false") {
     forAll { (a: Doc, b: Doc) =>
       import Doc._
-      val atree = toDocTree(a)
-      val btree = toDocTree(b)
-      if (isSubDoc(atree, btree)) {
-        setDiff(btree, atree) match {
+      val atree = DocTree.toDocTree(a)
+      val btree = DocTree.toDocTree(b)
+      if (DocTree.isSubDoc(atree, btree)) {
+        DocTree.setDiff(btree, atree) match {
           case None =>
             // If a is a subset of b, and b - a == empty, then a == b
             assert(a.compare(b) == 0)
           case Some(diff) =>
-            assert(!isSubDoc(atree, diff))
+            assert(!DocTree.isSubDoc(atree, diff))
         }
       }
       else {
         /*
          * We either have disjoint, overlapping, or btree is a strict subset of atree
          */
-        setDiff(btree, atree) match {
+        DocTree.setDiff(btree, atree) match {
           case None =>
             // if we btree is a strict subset of of atree
-            assert(isSubDoc(btree, atree))
+            assert(DocTree.isSubDoc(btree, atree))
           case Some(bMinusA) =>
             // disjoint or overlapping, so atree and bMinusA are disjoint
-            assert(!isSubDoc(atree, bMinusA))
-            assert(((deunioned(atree).toSet) & (deunioned(bMinusA).toSet)).isEmpty)
+            assert(!DocTree.isSubDoc(atree, bMinusA))
+            assert(((DocTree.deunioned(atree).toSet) & (DocTree.deunioned(bMinusA).toSet)).isEmpty)
         }
       }
     }
@@ -305,7 +302,7 @@ the spaces""")
     forAll { (a: Doc, b: Doc) =>
       import Doc._
 
-      if (deunioned(a).toSet.subsetOf(deunioned(b).toSet)) {
+      if (a.deunioned.toSet.subsetOf(b.deunioned.toSet)) {
         assert(a.isSubDocOf(b))
       }
       // due to normalization, the other case may tell us nothing
